Moved windowsinfo.py from its unnecessary folder
